--- a/test/openjd/model/benchmark/__init__.py
+++ b/test/openjd/model/benchmark/__init__.py
@@ -1,1 +1,7 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+
+"""
+Benchmark tests for OpenJD model performance testing.
+
+This package contains performance benchmarks for various components of the OpenJD model.
+"""