### What was the problem/requirement? (What/Why)
Our yaml parser was far slower than our JSON parser which prevented certain checks we'd like to do at job creation time in some cases and generally caused our create job API to run slower than we'd like.

### What was the solution? (How)
Try to use the yaml CSafeLoader where available.  Create benchmarking tests to note the difference.

### What is the impact of this change?

```
### 1. Small Template (987 characters):
• **SafeLoader**: 2.91ms
• **CSafeLoader**: 0.27ms
• **Performance improvement**: 11.0x faster
• **Time saved**: 2.64ms per parse

### 2. Large Template (44,187 characters):
• **SafeLoader**: 125.74ms
• **CSafeLoader**: 13.44ms
• **Performance improvement**: 9.4x faster
• **Time saved**: 112.30ms per parse

### 3. Extra Large Template (192,387 characters):
• **SafeLoader**: 590.07ms
• **CSafeLoader**: 65.48ms
• **Performance improvement**: 9.0x faster
• **Time saved**: 524.59ms per parse

### 4. File-based Template (27,347 characters):
• **SafeLoader**: 77.26ms
• **CSafeLoader**: 8.08ms
• **Performance improvement**: 9.6x faster
• **Time saved**: 69.18ms per parse
```

### How was this change tested?
All unit tests pass
Benchmarking test created which passes and shows above performance improvements

### Was this change documented?
Yes

### Is this a breaking change?
No

### Does this change impact security?
No - the CSafeLoader is a C implementation of the safe loader we were using previously - https://pyyaml.org/wiki/PyYAMLDocumentation

----

*By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.*